refactor: rename download service to install and update imports
